requirements.txt
ultralytics
cv2
schedule
os
pyyaml